requirements.txt
Django==2.2.2
djangorestframework==3.9.4
request==2019.4.13
requests==2.22.0
reportlab==3.5.23
urllib3==1.25.3